+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This file is used to list changes made in each version of cookbook-openstack-com
 * Add a new custom matcher render_config_file to test ini file content
 * Add global rabbit ha flag
 * Removed all ceph logic
+* Allow different URLs for internal, public, admin endpoints (bug 1412919)
 
 ## 10.1.0
 * Adding identity admin bind host endpoint to allow flexibility and consistency